+++ Snake_7/food.cpp
@@ -4,7 +4,7 @@
 #include <ctime>
 
 
-void generate_food(int& m_food_X, int& m_food_Y, bool& m_flag, char(*m_field)[rows]) {
+void generate_food(size_t& m_food_X, size_t& m_food_Y, bool& m_flag, char(*m_field)[rows]) {
 	srand(time(0));
 	while (!m_flag) {
 		m_food_X = rand() % visual_max_X + 1;
