+++ Snake_7/food.h
@@ -1,3 +1,3 @@
 #pragma once
 
-void generate_food(int& m_food_X, int& m_food_Y, bool& m_flag, char(*m_field)[rows]);
+void generate_food(size_t& m_food_X, size_t& m_food_Y, bool& m_flag, char(*m_field)[rows]);
